Refactored clientProvider, added metrics reporter
