+++ worker-queue-sqs/src/main/java/com/hpe/caf/worker/queue/sqs/util/SQSUtil.java
